(maint) Update beaker to use newest puppet repos
This commit updates beaker to use the new puppet5 repos. However, the
repos that this commit updates beaker to use are the rolling repos.
These are links to the currently active major version series. When
puppet6 is out and stable, we will switch these puppet repos to link to
the puppet6 repos. We will update these links for every new major
version series. This means that we will not have to make any changes in
beaker unless we want to pin a project to a particular version series,
i.e., 5. Ideally we wouldn't need to pin, but that's not super realistic