(BKR-1159) Reset opts in install_puppet_agent_dev_repo_on
This commit moves the setting of the values for `opts` into the block
for the hosts submitted to the `install_puppet_agent_dev_repo_on`
method. Prior to this commit, the `opts[:download_url] would become
corrupted by inheriting the value from the previous host run and
appending to it.