(BKR-1343) Install as an MSI on Windows
On Windows, the "install_from_build_data_url" method was trying to
install the artifact as a cygwin package which won't work for MSIs.
Instead use the beaker method for installing as an MSI. Note the
"install_msi_on" method is specific to the puppet-agent package, because
it will ensure the puppet service is stopped before installing and will
install it as a Manual service. In the future we may want to
differentiate between the puppet-agent project and others, and use
"generic_install_msi_on" for the latter.