(maint) Don't join with URI for constructing urls
If we have a base uri that doesn't end with a slash, URI.join will just
cut off that part of the uri, which is very unexpected.

For example, if we call URI.join('http://nightlies.puppet.com/apt',
'puppet5-nightly-release-xenial.deb'), we will end up with `#<URI::HTTP
http://nightlies.puppet.com/puppet5-nightly-release-xenial.deb>`, which
isn't what we want.

This commit simply remote the URI.join logic in favor of what we're
already doing for rpm release packages. This simply constructs the URI
as a string, and doesn't take advantage of the uri library at all.