(BKR-1443) Ensure we do not destroy hosts if specified
Prior to this commit, when running acceptance tests from the rakefile,
we were always destroying the host if the tests passed. However, this
didn't take into account when users passed in `--preserve-hosts=always`.
This commit ensures we skip the `beaker destroy` step whenever the user
sets `--preserve-hosts=always` in the OPTIONS setting.