(maint) Use generic install msi method
`install_msi_on` includes a lot of puppet specific utilities when
installing the puppet-agent package, including a check to the state of
the puppet service to make sure the MSI setup went as expected. There
are also a few registry checks to ensure the puppet MSI installation
went as expected. These checks should be moved to the puppet acceptance,
not in beaker-puppet. This commit uses the more basic
`generic_install_msi_on` method, which will work for installing MSIs
that are not puppet-agent, like bolt or PDK.