(maint) Use --module-repository for a custom forge
The current beaker workflow with a custom forge host means stubbing it
in /etc/hosts. This doesn't work because it uses SSL and then the
certificate validation fails. By using the --module-repository option it
just works and even allows you to use plain http and a different port.

Tested with https://github.com/unibet/puppet-forge-server to use caching
to avoid the slow puppet forge.