(maint) Use install_msi_on for puppet-agent
This is a partial revert of commit 5a9c90fa187d. When installing
puppet-agent packages we need to use the install_msi_on helper, because
it sets the service startup type to manual, which prevents the service
from starting and sending a CSR before we've configured the agent. It
also verifies that the service startup is correct after installation,
otherwise you can get transient failures if the background agent service
is running while performing puppet agent -t runs in the foreground.

When installing MSI packages for other projects, use the
generic_install_msi_on method.