(maint) Simplify logic for skipping agent that is also a master
When the master host also has an agent role, we don't want to manually
run `puppet agent -t`, because puppetserver will generate the
self-signed CA and server certs when it starts. Rather than checking for
the role, enumerate the agents, skipping the agent if it is also the
master host.

Add a comment about why we're deleting the contents of the ssldir
instead of the directory.