(PUP-8305) Git testing should consume the runtime package
This commit includes a lot of changes that enable git-based acceptance
testing. First and foremost, we're satisfying runtime dependencies with
the new puppet-runtime tarball. This provides a host of third party
packages, namely ruby.

Once this is in place, we can remove quite a bit of unnecessary logic
that was previously required when using system-provided packages (or
hand built packages).

This also changes which defaults are loaded for git-based testing.
Previously, we used the foss defaults. We need to switch to using the
same modern paths that are set with the aio defaults.