(PE-25146) Check for hostnames when puppetserver signing
We were seeing race conditions in ci for frictionless installs on
solaris and ubuntu (slower platforms?) where the frictionless install
script starts the puppet service, which runs the agent, which begins the
csr submission.

The 'puppetserver ca list --all' output now returns 'Requested' certs in one
section and 'Signed' certs in another. The test for a lack of
'Requested' section works so long as we haven't gotten here before the
agent service has completed submitting the CSR.

To work around this, verify that hostnames are all signed by checking
that we have no Requested, and that therefore we're finding hostnames in
the Signed section.