(PUP-9136) Ensure state is preserved between runs
We need to be able to preserve the state of the hosts that is set up
during the pre-suite runs. We previously were relying on beaker to
initiate this on every invocation, but that behavior was removed in
beaker 4. This commit takes advantage of a new flag that will cause
invocations of `beaker exec` to save the host state. We really only want
to do this when we're running pre-suites.

The `--preserve-state` flag was introduced in beaker 4.1.0. If we try to
run these tasks with a lower version of beaker, we'll get a failure
because of an unknown option.