(PUP-8305) Git testing should consume the runtime package (#52)
* (PUP-8305) Git testing should consume the runtime package

This commit includes a lot of changes that enable git-based acceptance
testing. First and foremost, we're satisfying runtime dependencies with
the new puppet-runtime tarball. This provides a host of third party
packages, namely ruby.

Once this is in place, we can remove quite a bit of unnecessary logic
that was previously required when using system-provided packages (or
hand built packages).

This also changes which defaults are loaded for git-based testing.
Previously, we used the foss defaults. We need to switch to using the
same modern paths that are set with the aio defaults.

* (maint) We do not always have @options set when running commands

* (maint) Cleanup type regex

* (maint) Ensure we get the tag from this branch

Apparently there are some potential issues with `git describe` where it
will potentially pick up the "older" branch because it happens "first"
in the history (i.e., from a merge-up). This will ensure we stay on the
branch we have checked out.

* updates

* Setup git-based projects only on agents

* update setup task paths

* (maint) Install windows into the expected path

* remove pry

* (maint) Ensure state is preserved

* (maint) Go back to finding the latest runtime tag

* wip

* windows path changes

gotta support both puppet 5.5.x and puppet 6 :(

* (maint) remove byebug call

* do we need these lines?

* (maint) Remove calls to install.rb

* we shouldn't need this anymore

* Revert "(maint) Remove calls to install.rb"

This reverts commit 2add620e50b4091ef5ff7f7921d87356f286da3c.

* Revert "Revert "(maint) Remove calls to install.rb""

This reverts commit 524601e8cd486e4f87eef5b939f8a8ea04f5d118.

* add RUBYLIB so we can load puppet via ruby/irb