(maint) Distinguish between ruby_arch x64 and x86
Rather than throwing all the possible windows paths with both `Program
Files (x86)` and `Program Files`, we should distinguish between which
we expect for the system we are working on. This will ensure that we
install puppet into the expected location depending on what `ruby_arch`
is set to. This commit also adds the new pathing for ruby that was added
for puppet 6. We need to keep the old path as well for backwards
compatibility.