(PUP-8305) Simplify installing gems from git repo
For testing purposes, we add a reference to the repo we want to test to
a Gemfile, then install it with `bundle install`. This commit removes
some previously required branching now that we are using ruby from the
puppet-runtime archive.

I am also removing any calls to `install.rb` here. We use `install.rb`
to move files into place, and so some basic setup. However, that
shouldn't be required when we use bundler. There are likely some setup
steps that were happening as a part of `install.rb` that will need to
be moved to project specific pre_suite tests (i.e., registering
puppetres.dll for puppet on windows), but those pre_suites should live
in the corresponding project repo.

We also only want to run 'bundle install' on the agent hosts, as we are
installing puppetserver from packages on the master server, and so we
need to resolve the puppet-agent dependency with the puppet-agent
package. This isn't perfect, as a lot of our testing defines the master
server with both master and agent roles, but we might as well try.