(BKR-1558) make sure we destroy hosts for ci:test:* task execution
When running any of the ci:test:* test tasks, the sequence of beaker
subcommands to run is defined by beaker_suite, which in turn uses
beaker and then sh. The sh command will exit with a status of 1 when
tests fail. This in turn aborts rake, which then prevents any of the
commands following test execution to be run.

In this specific instance, when tests fail, we won't run post-suite
or de-provision machines via destroy. The end result is that when
tests fail, we leave machines provisioned until they are reaped.

This change ensures that beaker destroys hosts as it should, regardless
of running via the mainline flow or with subcommands.