(maint) `agent` should be `host`
It turns out `agent` and `agents` are equivalent if you don't explicitly set
`agent` to something. Who knew! The loop that this exists in we refer to
the individual servers as `host`, not `agent` as I had previously, and
incorrectly, assumed.

This resulted in some strange errors depending on how many agents we had
defined for a given run. `agent`, and I assume `agents` as well, returns
a host object if there is only one host with the agent role and it will
return an array of hosts if there are multiple hosts with the agent
role.