(maint) Remove assumption that we have runtime on docker
It's too much to assume that we have a docker image that has all the
runtime bits already installed. We should remove this assumption. This
will allow us to run both package based and git based acceptance tests
on docker machines. If and when we have a docker image pipeline that
provides regularly updated docker images with runtime on them already,
we can change this back.

In the meantime, the automation here is still dependent on resources
that are only available on infrastructure internal to Puppet, Inc. This
means that running acceptance tests for our community members is still
not easy.