(maint) Restructure release repo url
This change is prompted by this email:
https://groups.google.com/d/msg/puppet-users/cCsGWKunBe4/OdG0T7LeDAAJ

As a summary, the layout of yum.puppet.com and nightlies.puppet.com/yum
are changing slightly. The Release Engineering team is cleaning up the
layout of these two repos, specifically where release packages are
symlinked. This commit updates where we anticipate finding the release
packages in accordance with the changes outlined in the above email.