(maint) Rewrite agent/server install logic
We had, and still have, a ton of duplication of logic. This commit
attempts to clean up a bit of it.

In the rake tasks used for test setup, a lot of code/logic is moved out
of the rake task and into ruby classes. This ensures that this code is
actually testable.

I'm also reducing some diverging behavior between the
`install_puppet_agent_on` and the `install_puppetserver_on` methods.
Ideally, these two would use one main package install method that has
the same ability to hierarchically pull packages.

That hierarchy ought to look something like:
builds.d.pl.net -> nightlies.p.com -> {yum|apt|downloads}.p.com

There's still a ton of work I'd like to see happen with this codebase,
including a significant test overhaul, but this is a little bit of
streamlining to at least start the project.