(maint) Add extra customization to the nightly repo urls
We have the ability to either specify a main custom nightly repo url, or
to supply custom apt/yum nightly repo urls. If the main one is set to a
custom value, but apt and yum are not, then use the main for both apt
and yum. If the main value is set to custom, and apt and yum are also
set to custom, do not overwrite the custom apt and yum values.