(maint) Mark dev repos as trusted
Instead of globally disabling package signing checks we can mark our development
repos as trusted. This allows both the repository and its packages to be
unsigned.

The previous implementation was more broad and more specific than needed, in
that it applied to all repositories and required separate settings for unsigned
repositories and unsigned packages. The setting for unsigned packages was
missing, which caused failures on the platforms that required it.

On older versions where this flag is not supported it is silently
ignored. It is supported in all versions that check for signed
repositories or packages.