(BKR-1600) Default puppet settings to `main` section
If puppet.conf contained a setting without a preceding section, as can happen
due to PUP-4755, then the 'lay_down_new_puppet_conf' method created a 'global'
section for that setting, however, that is not a valid puppet section name.

Specify the default section name as 'main' so it matches puppet's behavior.