(RE-12690) Add debug output to `install_repo_configs_from_url`
This commit adds debug output for the repo config contents before and after
gsubbing. Filesync tests have been experiencing a transient failure where
debian repo contents are malformed. We've narrowed down the failure to this
method, somewhere between pulling down the file and scp'ing it to the host.
This debug output is intended to give us more insight into what is causing this
issue, but I haven't actually been able to reproduce the issue since making
this change. Perhaps this is enough to eliminate the transient altogether.
¯\_(ツ)_/¯