Provide a useful message if possible ISP hijacking
I stumbled across this with an ISP that was hijacking failed web
connections and returning valid HTTP pages. This was causing the code to
think that it was on the puppet development network and, of course,
fail.

I also now clean up the temp directory for downloaded YAML artifacts in
Puppet 5.

This needs to be reproduced across all of the code and handled properly.