(maint) Remove puppet teams from CODEOWNERS file
We are planning to transfer this repo to the voxpupuli organization, and
once it's not in the Puppet organization Puppet teams can't be
referenced in the file. This removes Puppet teams from the file, with
the option to replace individual usernames for those interested in
continuing to maintain the repo.