(maint) Make "latest" install latest stable instead of nightly
Prior to this commit, if `install_puppetserver_on` was called without a
version it would install the latest available nightly build.

This logic makes the use case of installing the latest *released*
version unusable (i.e. from non-nightly collections), as the method
would always append `-nightly` to the collection if no version is
passed.

As the `nightly` option can already be supplied to the method, change
the logic to no longer assume the user wants a nightly build if they
don't pass a version (or if `version: 'latest'`).