Create better puppet apply tmpfiles
Adds a date-based incremental portion to the /tmp/apply_manifest*.pp files so
that debugging is easier.

Prevously, you had a set of files that were not ordered by application time.
Now, you have a set of files that is ordered by application time which means
that a simple directory listing of /tmp will tell you which one executed last.