(PA-4331) Add a step for updating Solaris CA Keystore
A recent change to the way puppet-agent is packaged requires versions
of Solaris that are older then 11.3 to install a new cert.
The Puppet VM images for Solaris for 11, 11.2, and Solaris-Sparc need to
be updated to include the updating of the keystore. That will take some
time to accomplish, as a work around we are going to update the keystore
as part of a pre-suite.
Once we update our vmpooler images for the three versions we can remove
the pre-suite step.
Any other OS will skip the pre-suite step.