(PA-4178) Only clear and add non-default gem sources
Rubygems 3.1 and up will prompt when adding a gem source that is too similar to
rubygems.org to protect against typo squatting. This causes tests to fail if
`GEM_SOURCE` is not overridden:

    gem source --add https://rubygems.org
      ERROR:  While executing gem ... (Gem::OperationNotSupportedError)
        Not connected to a tty and no default specified

Since rubygems always includes https://rubygems.org as the default source, we
now only clear and add the source when it's not the default.

This issue was introduced due to changes in rubygems/bundler to prevent typo
squatting[1]. However, it is also triggered if the source matches `rubygems.org`
exactly.

[1] https://github.com/rubygems/rubygems/commit/aa967b85dd96bbfb350f104125f23d617e82a00a