Expect commands to be executed on hosts
Previously the tests expected Beaker::Command objects to be created, but not
that those commands were ever executed on a host. And since the commands were
created for both hosts, the test did a lot of `.exactly(2).times`

This updates the tests to verify that on(host, ..) is called with either the
expected string or an object whose `command` attribute has the expected value.

It also adds a test for the 'PUPPET_AGENT_STARTUP_MODE => Disabled' case.