(PA-4552) Create ci:test:setup step for iterative workflow
It's now possible to setup host(s) in preparation for puppet agent related
beaker tests using:

    SHA=<SHA> HOSTS=redhat7-64a bundle exec rake ci:test:setup

SHA may be a tag or a full 40 digit git SHA.

HOSTS may be specified as a beaker-hostgenerator string or a file to an existing
hosts file. If not a file, then the abs hypervisor will be used, similar to the
`ci:test:aio` task.

For facter and puppet-agent tests, HOSTS defaults to a single redhat7-64a host.
For pxp-agent and puppet tests, HOSTS defaults to a redhat7-64m server and
redhat7-64a agent. The `SERVER_VERSION` environment variable can be used to
select a specific puppetserver version, otherwise the last released version will
be used.

When setup is complete, run 'bundle exec beaker exec path/to/test.rb' to test
iteratively.

There is a bug in beaker-abs that prevents hosts from being destroyed via
`bundle exec beaker destroy`. They must be destroyed manually for now using
floaty.