(PA-4566) Retry beaker exec if presuite completes
Previously, if the beaker-puppet presuite or project-specific presuite
steps failed, then beaker_retry_suite would try to get a list of failed tests
and fail with:

    NoMethodError: undefined method `length' for nil:NilClass

This was because the test results file was empty, so tests_to_rerun was nil.

Now we only attempt to retry "beaker exec" if the presuite steps finished and
the list of failed tests is not empty.

If the list of failed tests is empty, then we reraise the exception. This seems
safer than blindly retrying, not knowing if beaker didn't run any tests or it
ran *all* of them and failed to write out the test results file.