Always run post-suite if provision succeeded
The `ci:test:aio` rake task would skip the `post-suite` if any presuites or
tests failed. Now the `post-suite` is protected in a nested ensure block so if
`provision` succeeds, then the `post-suite` is guaranteed to be called. This is
important in order to capture failures when installing the agent during the
`pre-suite`.

And due to the nested ensure, if the `post-suite` fails, then `destroy` will
also be guaranteed to be called.

```
$ export BEAKER_PUPPET_VERSION=/home/josh/work/beaker-puppet
$ bundle update
$ bundle exec rake 'ci:test:aio[true]' SHA=8.4.0 TESTS=tests/load_libfacter.rb HOSTS=hosts.yaml
...
Begin tests/load_libfacter.rb

C100161: Ruby can load libfacter without raising an error
RuntimeError: Whoops
/home/josh/work/facter/acceptance/tests/load_libfacter.rb:5
...
beaker exec post-suite
...
Begin teardown/common/099_Archive_Logs.rb
```