Add the ability to install puppet-agent from a local dev build
If the `DEV_BUILD_PATH` environment variable is set and refers to a
puppet-agent package, then it will be scp'ed to the remote system and
installed. This makes it possible to run beaker tests on a platform
without needing to generate repos or ship artifacts.

    export DEV_BUILD_PATH=/home/josh/work/puppet-agent/output/sles/12/puppet8/x86_64/puppet-agent-8.3.1.65.gc65ba7001-1.sles12.x86_64.rpm
    bundle exec rake ci:test:setup SHA=8.3.1 HOSTS=sles12-64a
    bundle exec beaker exec <test>

Where <test> is the file or directory of tests to execute against the remote system.

The SHA parameter is required, but unused.